Engine: Set CMake miminum version required to 3.14 and enable static CRT on windows.

+++ engine/Graphics/D3D12/D3D12DescriptorHeap.cpp
@@ -61,7 +61,7 @@ namespace Alimer
         }
 
         D3D12_CPU_DESCRIPTOR_HANDLE ret = currentHandle;
-        currentHandle.ptr += count * descriptorSize;
+        currentHandle.ptr += (SIZE_T)count * descriptorSize;
         remainingFreeHandles -= count;
         return ret;
     }
